[system_health] move get_skip_mod_list to tests/common to fix cross-feature import
The Pre_test Dependency Check rejects imports across feature directories.
Move get_skip_mod_list from tests/platform_tests/cli/util into
tests/common/platform/device_utils, which is the existing home for shared
platform helpers. Re-export it from the original location so existing callers
in platform_tests/ are unaffected without any changes on their end.
