+++ tests/platform_tests/cli/util.py
@@ -4,6 +4,7 @@ util.py
 Utility functions for testing SONiC CLI
 """
 import six
+from tests.common.platform.device_utils import get_skip_mod_list  # noqa: F401
 
 
 def parse_colon_speparated_lines(lines):
@@ -66,43 +67,6 @@ def get_fields(line, field_ranges):
         fields.append(field.strip())
 
     return fields
-
-
-def get_skip_mod_list(duthost, mod_key=None):
-    """
-    @summary: utility function returns list of modules / peripherals absent in chassis
-    by default if no keyword passed it will return all from inventory file
-    provides a list under skip_modules: in inventory file for each dut
-    returns a empty list if skip_modules not defined under host in inventory
-    inventory example:
-    DUTHOST:
-    skip_modules:
-        'line-cards':
-          - LINE-CARD0
-          - LINE-CARD2
-        'fabric-cards':
-          - FABRIC-CARD3
-        'psus':
-          - PSU4
-          - PSU5
-        'thermals':
-          - TEMPERATURE_INFO_2
-    @return a list of modules/peripherals to be skipped in check for platform test
-    """
-
-    skip_mod_list = []
-    dut_vars = duthost.host.options['variable_manager'].get_vars()['hostvars'][duthost.hostname]
-    if 'skip_modules' in dut_vars:
-        if mod_key is None:
-            for mod_type in list(dut_vars['skip_modules'].keys()):
-                for mod_id in dut_vars['skip_modules'][mod_type]:
-                    skip_mod_list.append(mod_id)
-        else:
-            for mod_type in mod_key:
-                if mod_type in list(dut_vars['skip_modules'].keys()):
-                    for mod_id in dut_vars['skip_modules'][mod_type]:
-                        skip_mod_list.append(mod_id)
-    return skip_mod_list
 
 
 def get_skip_logical_module_list(duthost, mod_key=None):
